Split backend tests by behavior
